==> step1.LoadCleanExplore.R <==
######Load packages#####
library(lmSupport)
library(psych)
library(dplyr)
library(ggplot2)
library(performance)
library(data.table)
library(foreign)
library(tidyverse)
library(lavaan)
###maybe need to load but not every time
library(tidymodels)
library(doParallel)
library(pROC)
rm(list = ls())
gc()

####------------PNC-------------####

###note, this version of the data only includes AA and EA; mixed race individuals were not included
PNCFullSample = readxl::read_xlsx("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/PNC projects/Coding and analysis/Diagnosis result file/Processed data file/PNC Diagnoses_withAge.xlsx", na = "-999")
PNCfactorScore = readxl::read_xlsx("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNCBiThree_FactorScore_xlsxFormat.xlsx", na="-999")
PNCfactorScore <- PNCfactorScore %>%
  select(SUBJID, G, INT, EXT, THOUGHT)
PC = paste("PC", seq(1,10, by = 1), sep = "")


##Load PGSs###

####GSEM ADHD -> don't run; not relevant for this project
#pnc_EAADHDPGSGSEM = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/CHGPM Grant Study (ASD_ADHD)/GSEM Specific Association Results/GSEM_PGS_0516_noSUB/PNC/PNC_EA_GSEMADHD_05162021_noSUB.all_score",sep="",header = TRUE) %>%
#  rename(pgs_gsemADHD = Pt_1)
#pnc_AA_ADHDPGSGSEM = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_GSEMADHD_noSUB.all_score",sep="",header = TRUE) %>%
#  rename(pgs_gsemADHD = Pt_1)
#pnc_allgsemADHD <- rbind(pnc_EAADHDPGSGSEM,pnc_AA_ADHDPGSGSEM)

####replicated ADHD -> replicated because trad ADHD was almost the only PGS that had significant and robust association
#####the previous ADHD PGS might have been generated by Q's lab, so I want to check if I can generate similarly powerful PGS using my own script and protocol
#pnc_EAADHDPGSrepl = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ADHDtrad_10172021.all_score",sep = "",header = TRUE) %>%
#  rename(pgs_adhdrepl = Pt_1)
#pnc_AAADHDPGSrepl = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ADHDtrad_10172021.all_score",sep = "",header = TRUE) %>%
#  rename(pgs_adhdrepl = Pt_1)
#pnc_alladhdreplpgs <- rbind(pnc_EAADHDPGSrepl, pnc_AAADHDPGSrepl)

####SCZ
#pnc_EASCZPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_SCZ.all.score",sep = "",header = TRUE) %>%
#  rename(pgs_scz = X1)
#pnc_AASCZPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_SCZ.all.score",sep = "",header = TRUE) %>%
#  rename(pgs_scz = X1)
#pnc_allsczpgs <- rbind(pnc_EASCZPGS, pnc_AASCZPGS)

####ANX  ->> as of 12/3/2021, using GSEM ANX created by Zihang
#pnc_EAANXPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ANXgsem.all_score",sep = "",header = TRUE) %>%
#  rename(pgs_anx = Pt_1)
#pnc_AAANXPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ANXgsem.all_score",sep = "",header = TRUE) %>%
#  rename(pgs_anx = Pt_1)
#pnc_allanxpgs <- rbind(pnc_EAANXPGS, pnc_AAANXPGS)

####MDD -> using rerun PGS after changing PRSice-2 flag from -or to -beta. The GWAS effect was in LogOR
#pnc_EAMDDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_MDD_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_mdd = Pt_1)
#pnc_AAMDDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_MDD_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_mdd = Pt_1)
#pnc_allMDDPGS <- rbind(pnc_EAMDDPGS, pnc_AAMDDPGS)

###BIP -> using rerun PGS (beta flag)
#pnc_EABIPPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_BIP_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_bip = Pt_1)
#pnc_AABIPPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_BIP_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_bip = Pt_1)
#pnc_allBIPPGS <- rbind(pnc_EABIPPGS, pnc_AABIPPGS)

###PTSD EUR -> replacing with GSEM rerun PTSD generated by zihang. European sample
#pnc_EAPTSDallPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_PTSDgsem.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdgsem = Pt_1)
#pnc_AAPTSDallPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_PTSDgsem.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdgsem = Pt_1)
#pnc_allPTSDallPGS <- rbind(pnc_EAPTSDallPGS, pnc_AAPTSDallPGS)

###PTSD EUR 
#pnc_EAPTSDeurPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_PTSD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdeur = X1)
#pnc_AAPTSDeurPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_PTSD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdeur = X1)
#pnc_allPTSDeurPGS <- rbind(pnc_EAPTSDeurPGS, pnc_AAPTSDeurPGS)

###PTSD AFR
#pnc_EAPTSDafrPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_PTSD_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdafr = X1)
#pnc_AAPTSDafrPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_PTSD_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdafr = X1)
#pnc_allPTSDafrPGS <- rbind(pnc_EAPTSDafrPGS, pnc_AAPTSDafrPGS)

###AN
#pnc_EAANPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_AN.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_an = X1)
#pnc_AAANPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_AN.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_an = X1)
#pnc_allANPGS <- rbind(pnc_EAANPGS, pnc_AAANPGS)

###OCD
#pnc_EAOCDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_OCD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ocd = X1)
#pnc_AAOCDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_OCD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ocd = X1)
#pnc_allOCDPGS <- rbind(pnc_EAOCDPGS, pnc_AAOCDPGS)

###ASD GWAS
#pnc_EAASDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ASD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_asd = X1)
#pnc_AAASDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ASD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_asd = X1)
#pnc_allasdPGS <- rbind(pnc_EAASDPGS, pnc_AAASDPGS)

###TS GWAS
#pnc_EATSPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_TS.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ts = X1)
#pnc_AATSPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_TS.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ts = X1)
#pnc_alltsPGS <- rbind(pnc_EATSPGS, pnc_AATSPGS)

###My GSEM GWAS - merged; with substance; four factors: internalizing, substance, ndd and tp
#pnc_EAGSEMMergedPGS = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/gsem_pgs_fivefactor_02222022/EUR_merged_imputed_qced_No_Axiom_EUR.pgs.merged") %>%
#  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
#         pgs_ndd_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
#         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
#         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
#         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)
#pnc_AAGSEMMergedPGS = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/gsem_pgs_fivefactor_02222022/AA_merged_imputed_qced.pgs.merged") %>%
#  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
#         pgs_ndd_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
#         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
#         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
#         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)
#pnc_all_gsemMergedPGS = rbind(pnc_EAGSEMMergedPGS,pnc_AAGSEMMergedPGS)

###GSEM GWAS - Compulsive - no substance
#pnc_EACompGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_CompulsiveGSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_comp_gsem = Pt_1)
#pnc_AACompGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_CompulsiveGSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_comp_gsem = Pt_1)
#pnc_allCompGSEMPGS <- rbind(pnc_EACompGSEMPGS, pnc_AACompGSEMPGS)

###GSEM GWAS - Neurodevelopment - no substance
#pnc_EANeuroGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_Neuro_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_neuro_gsem = Pt_1)
#pnc_AANeuroGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_Neuro_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_neuro_gsem = Pt_1)
#pnc_allNeuroGSEMPGS <- rbind(pnc_EANeuroGSEMPGS, pnc_AANeuroGSEMPGS)

###GSEM GWAS - Thought - no substance
#pnc_EATHTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_THT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_tht_gsem = Pt_1)
#pnc_AATHTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_THT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_tht_gsem = Pt_1)
#pnc_allTHTGSEMPGS <- rbind(pnc_EATHTGSEMPGS, pnc_AATHTGSEMPGS)

###GSEM GWAS - Internalizing - no substance
#pnc_EAINTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_INT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_int_gsem = Pt_1)
#pnc_AAINTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_INT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_int_gsem = Pt_1)
#pnc_allINTGSEMPGS <- rbind(pnc_EAINTGSEMPGS, pnc_AAINTGSEMPGS)


###AUDIT
#pnc_EAAUDITPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_AUDIT.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_audit = X1)
#pnc_AAAUDITPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_AUDIT.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_audit = X1)
#pnc_allauditPGS <- rbind(pnc_EAAUDITPGS, pnc_AAAUDITPGS)
###ALCH Depend full discovery
#pnc_EAALCHPGS_full = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ALCHdep_fulldisc.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchfull = X1)
#pnc_AAALCHPGS_full = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ALCHdep_fulldisc.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchfull = X1)
#pnc_allalchPGSfull <- rbind(pnc_EAALCHPGS_full, pnc_AAALCHPGS_full)

###ALCH Depend eur discovery
#pnc_EAALCHPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ALCHdep_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alcheur = X1)
#pnc_AAALCHPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ALCHdep_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alcheur = X1)
#pnc_allalchPGSeur <- rbind(pnc_EAALCHPGS_eur, pnc_AAALCHPGS_eur)

###ALCH Depend afr discovery
#pnc_EAALCHPGS_afr = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_ALCHdep_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchafr = X1)
#pnc_AAALCHPGS_afr = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_ALCHdep_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchafr = X1)
#pnc_allalchPGSafr <- rbind(pnc_EAALCHPGS_afr, pnc_AAALCHPGS_afr)

###CUD EUR GWAS
#pnc_EACUDPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_CUD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_cudeur = X1)
#pnc_AACUDPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_CUD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_cudeur = X1)
#pnc_allcudPGSeur <- rbind(pnc_EACUDPGS_eur, pnc_AACUDPGS_eur)

###Externalizing consortium GWAS
pnc_EAEXTconsPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_EA_Externalizing.all.score", sep = "", header = TRUE) %>%
  rename(pgs_extcons = X1)
pnc_AAEXTconsPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/PNC_AA_Externalizing.all.score", sep = "", header = TRUE) %>%
  rename(pgs_extcons = X1)
pnc_allEXTconsPGS <- rbind(pnc_EAEXTconsPGS, pnc_AAEXTconsPGS)

###Irwin - GSEM
pnc_EAIrwinPGS = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/irwin_gsem_EUR_merged_imputed_qced_No_Axiom_EUR.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_ext_irwin = pgs_from_ext_mvGWAS_finaledited.txt,
         pgs_int_irwin = pgs_from_int_mvGWAS_finalEdited.txt,
         pgs_ndd_irwin = pgs_from_ndd_mvGWAS_finalEdited.txt,
         pgs_tp_irwin = pgs_from_tp_mvGWAS_finalEdited.txt)
pnc_AAIrwinPGS = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/irwin_gsem_AA_merged_imputed_qced.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_ext_irwin = pgs_from_ext_mvGWAS_finaledited.txt,
         pgs_int_irwin = pgs_from_int_mvGWAS_finalEdited.txt,
         pgs_ndd_irwin = pgs_from_ndd_mvGWAS_finalEdited.txt,
         pgs_tp_irwin = pgs_from_tp_mvGWAS_finalEdited.txt)
pnc_allIrwinPGS <- rbind(pnc_EAIrwinPGS, pnc_AAIrwinPGS)

###individual disorder PGSs 
pnc_EA_indDisorder = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/individualDisorder_EUR_merged_imputed_qced_No_Axiom_EUR.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_alch = pgs_from_ALCH_gsem_clean.txt.gz,
         pgs_anx = pgs_from_ANX_gsem_clean.txt.gz,
         pgs_cud =pgs_from_CUD_EUR_casecontrol_public_11.14.2020.gz,
         pgs_adhd = pgs_from_daner_adhd_meta_filtered_NA_iPSYCH23_PGC11_sigPCs_woSEX_2ell6sd_EUR_Neff_70.meta.gz,
         pgs_asd = `pgs_from_iPSYCH-PGC_ASD_Nov2017.gz`,
         pgs_ocd = pgs_from_ocd_aug2017.gz,
         pgs_mdd = `pgs_from_PGC_UKB_depression_genome-wide.txt`,
         pgs_bip = `pgs_from_pgc-bip2021-all.vcf.tsv`,
         pgs_scz = pgs_from_PGC3_SCZ_wave3_public.v2.tsv.gz,
         pgs_an = `pgs_from_pgsAN2.2019-07.reordered.vcf.tsv`,
         pgs_ptsd = pgs_from_PTSD_gsem_clean.txt.gz,
         pgs_smkinit = pgs_from_SmokingInitiation.txt,
         pgs_ts = pgs_from_TS_Oct2018.gz)
pnc_AA_indDisorder = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/individualDisorder_AA_merged_imputed_qced.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_alch = pgs_from_ALCH_gsem_clean.txt.gz,
         pgs_anx = pgs_from_ANX_gsem_clean.txt.gz,
         pgs_cud =pgs_from_CUD_EUR_casecontrol_public_11.14.2020.gz,
         pgs_adhd = pgs_from_daner_adhd_meta_filtered_NA_iPSYCH23_PGC11_sigPCs_woSEX_2ell6sd_EUR_Neff_70.meta.gz,
         pgs_asd = `pgs_from_iPSYCH-PGC_ASD_Nov2017.gz`,
         pgs_ocd = pgs_from_ocd_aug2017.gz,
         pgs_mdd = `pgs_from_PGC_UKB_depression_genome-wide.txt`,
         pgs_bip = `pgs_from_pgc-bip2021-all.vcf.tsv`,
         pgs_scz = pgs_from_PGC3_SCZ_wave3_public.v2.tsv.gz,
         pgs_an = `pgs_from_pgsAN2.2019-07.reordered.vcf.tsv`,
         pgs_ptsd = pgs_from_PTSD_gsem_clean.txt.gz,
         pgs_smkinit = pgs_from_SmokingInitiation.txt,
         pgs_ts = pgs_from_TS_Oct2018.gz)
pnc_all_indDisorder <- rbind(pnc_EA_indDisorder, pnc_AA_indDisorder)

###GSEM new version with substances - merged, five factor - EUR reference
pnc_gsem_mergedPGS_EUR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/gsem_fivefactor_EUR_merged_imputed_qced_No_Axiom_EUR.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
         pgs_neuro_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)
###GSEM new version with substances - merged, five factor - EUR reference
pnc_gsem_mergedPGS_AFR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/PNC PGS/gsem_fivefactor_AA_merged_imputed_qced.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
         pgs_neuro_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)
pnc_all_gsem <- rbind(pnc_gsem_mergedPGS_EUR, pnc_gsem_mergedPGS_AFR)

###Merge all PGSs and change SUBJID
PGSCombined <- Reduce(function(...) merge(..., all = TRUE), 
                      list(pnc_all_gsem,pnc_all_indDisorder,
                           pnc_allEXTconsPGS,pnc_allIrwinPGS))
PGSCombined$SUBJID <- sapply(strsplit(PGSCombined$IID, split = "_", fixed = TRUE), function(x) (x[2]))
PGSCombined$FID = NULL
PGSCombined$IID = NULL
head(PGSCombined)
#add in traditional PGS from a previous file
TradADHDPGS = read.csv("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/R21 Resubmission 2021/PNC/PNC_AllPGS_AfterZihang_replicate.csv") %>%
  select(SUBJID, all_of(PC), ASD_EVER)
##there are duplicates -> probably because of parent report included in SUBJID as well. But PGSs are the same for the duplicate SUBJIDs; 
##remove all duplicates 
TradADHDPGS <- TradADHDPGS[-which(duplicated(TradADHDPGS$SUBJID)),]

###final data object with phenotype and genetic data
Merged_PNC <- merge(PNCFullSample, PGSCombined, by = "SUBJID")
Merged_PNC <- merge(Merged_PNC, TradADHDPGS, by = "SUBJID")
Merged_PNC <- merge(Merged_PNC, PNCfactorScore, by ="SUBJID")
table(Merged_PNC$Race)
Merged_PNC$age_at_cnb <- as.numeric(Merged_PNC$age_at_cnb)
Merged_PNC$Med_Rate <- as.numeric(Merged_PNC$Med_Rate)

###poisson models cannot allow negative values (G, INT, EXT and THOUGHT)
Merged_PNC$G <- Merged_PNC$G - min(Merged_PNC$G)
Merged_PNC$INT <- Merged_PNC$INT - min(Merged_PNC$INT)
Merged_PNC$EXT <- Merged_PNC$EXT - min(Merged_PNC$EXT)
Merged_PNC$THOUGHT <- Merged_PNC$THOUGHT - min(Merged_PNC$THOUGHT)

hist(log2(Merged_PNC$INT))
Merged_PNC <- Merged_PNC %>%
  mutate(G = log10(G+1),
         INT = log2(INT+1),
         EXT = log2(EXT+1))



###Variance of ADHD in PNC###
##variance of ADHD TSX
var(Merged_PNC$ADD_TSX)
var(Merged_PNC$ADD_TSX[Merged_PNC$Race==0])
var(Merged_PNC$ADD_TSX[Merged_PNC$Race==1])

varAncestry <- glm(ADD_TSX ~ Race + PC1 + PC2 + PC3 + PC4 + PC5 + PC6 +PC7 + PC8 + PC9 + PC10 +Sex + age_at_cnb, 
    family = "poisson", data = Merged_PNC)
summary(varAncestry)

##Basic Descriptive##
###Age and Sex by Race
##Sex: Male = 0, Female = 1, Other combinations = 2
##Race: AA = 0, EA = 1

table(Merged_PNC$Sex, Merged_PNC$Race)
SexR <- lm(Sex ~ Race,data = Merged_PNC)
anova(SexR)
  

####--------------------ABCD--------------------####
##load data
abcdPath <- "/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/ABCD/"
mergeCol <- c("subjectkey","src_subject_id","interview_date","interview_age","sex","collection_id")

###Demographics(parent demographics survey)
####don't run; will run a larger demographic object later
#abcd_pdem <- fread(paste0(abcdPath, "pdem02.txt")) %>%
#  slice(-1)%>%
#  filter(eventname == "baseline_year_1_arm_1") %>%
#  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))

##white (0=no, 1=yes)
#table(abcd_pdem$demo_race_a_p___10)
##black; African American
#table(abcd_pdem$demo_race_a_p___11)
##Native american
#table(abcd_pdem$demo_race_a_p___12)

##hispanic (1=yes; 2 = no)
#table(abcd_pdem$demo_ethn_v2)
##hispanic white
#table(abcd_pdem$demo_ethn_v2, abcd_pdem$demo_race_a_p___10)
#table(abcd_pdem$demo_ethn_v2, abcd_pdem$demo_race_a_p___11)

##chinese 
#table(abcd_pdem$demo_race_a_p___19)

####Genetic PCs
abcd_PCs <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Genetic_processing/ABCD/PCs_abcd.eigenvec")

###PGSs
####ADHD 
#abcd_ADHDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ADHD.all.score",sep = "",header = TRUE) %>%
#  rename(pgs_adhd = X1)

####SCZ
#abcd_SCZPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_SCZ.all.score",sep = "",header = TRUE) %>%
#  rename(pgs_scz = X1)

####ANX -> gsem version from zihang
#abcd_ANXPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ANXgsem.all_score",sep = "",header = TRUE) %>%
#  rename(pgs_anx = Pt_1)

####MDD -> rerun after using -beta flag because of LogOR
#abcd_MDDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_MDD_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_mdd = Pt_1)

###BIP
#abcd_BIPPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_BIP_rerun.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_bip = Pt_1)

###PTSD EUR GSEM
#abcd_PTSDallPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_PTSDgsem.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdgsem = Pt_1)
###PTSD EUR 
#abcd_PTSDeurPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_PTSD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdeur = X1)
###PTSD AFR
#abcd_PTSDafrPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_PTSD_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ptsdafr = X1)
###AN
#abcd_ANPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_AN.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_an = X1)
###OCD
#abcd_OCDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_OCD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ocd = X1)
###ASD GWAS
#abcd_ASDPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ASD.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_asd = X1)

###TS GWAS
#abcd_TSPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_TS.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_ts = X1)

###Externalizing consortium GWAS
abcd_EXTconsPGS_EUR = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_release_3.0_QCed_pgs_from_GSEM.GWAS.EXTERNALIZING.SHARE.v20191014.txt.gz.1kgCEU_maf0.01_rs_ref.all_score") %>%
  rename(pgs_extcons = pgs_from_GSEM.GWAS.EXTERNALIZING.SHARE.v20191014.txt.gz)
abcd_EXTconsPGS_AFR = read.delim("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_release_3.0_QCed_pgs_from_GSEM.GWAS.EXTERNALIZING.SHARE.v20191014.txt.gz.1000G_AFR_QC_rsID.all_score") %>%
  rename(pgs_extcons =  pgs_from_GSEM.GWAS.EXTERNALIZING.SHARE.v20191014.txt.gz)

###individual disorder PGS based on EUR reference
abcd_indDisorder_merged_EUR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/individualDisorder_ABCD_release_3.0_QCed.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_alch = pgs_from_ALCH_gsem_clean.txt.gz,
         pgs_anx = pgs_from_ANX_gsem_clean.txt.gz,
         pgs_cud =pgs_from_CUD_EUR_casecontrol_public_11.14.2020.gz,
         pgs_adhd = pgs_from_daner_adhd_meta_filtered_NA_iPSYCH23_PGC11_sigPCs_woSEX_2ell6sd_EUR_Neff_70.meta.gz,
         pgs_asd = `pgs_from_iPSYCH-PGC_ASD_Nov2017.gz`,
         pgs_ocd = pgs_from_ocd_aug2017.gz,
         pgs_mdd = `pgs_from_PGC_UKB_depression_genome-wide.txt`,
         pgs_bip = `pgs_from_pgc-bip2021-all.vcf.tsv`,
         pgs_scz = pgs_from_PGC3_SCZ_wave3_public.v2.tsv.gz,
         pgs_an = `pgs_from_pgsAN2.2019-07.reordered.vcf.tsv`,
         pgs_ptsd = pgs_from_PTSD_gsem_clean.txt.gz,
         pgs_smkinit = pgs_from_SmokingInitiation.txt,
         pgs_ts = pgs_from_TS_Oct2018.gz)
###individual disorder PGS based on AFR reference
abcd_indDisorder_merged_AFR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/individualDisorder_ABCD_release_3.0_QCed.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_alch = pgs_from_ALCH_gsem_clean.txt.gz,
         pgs_anx = pgs_from_ANX_gsem_clean.txt.gz,
         pgs_cud =pgs_from_CUD_EUR_casecontrol_public_11.14.2020.gz,
         pgs_adhd = pgs_from_daner_adhd_meta_filtered_NA_iPSYCH23_PGC11_sigPCs_woSEX_2ell6sd_EUR_Neff_70.meta.gz,
         pgs_asd = `pgs_from_iPSYCH-PGC_ASD_Nov2017.gz`,
         pgs_ocd = pgs_from_ocd_aug2017.gz,
         pgs_mdd = `pgs_from_PGC_UKB_depression_genome-wide.txt`,
         pgs_bip = `pgs_from_pgc-bip2021-all.vcf.tsv`,
         pgs_scz = pgs_from_PGC3_SCZ_wave3_public.v2.tsv.gz,
         pgs_an = `pgs_from_pgsAN2.2019-07.reordered.vcf.tsv`,
         pgs_ptsd = pgs_from_PTSD_gsem_clean.txt.gz,
         pgs_smkinit = pgs_from_SmokingInitiation.txt,
         pgs_ts = pgs_from_TS_Oct2018.gz)

###GSEM new version with substances - merged, five factor - EUR reference
abcd_gsem_mergedPGS_EUR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/gsem_fivefactor_ABCD_release_3.0_QCed.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
         pgs_neuro_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)
###GSEM new version with substances - merged, five factor - EUR reference
abcd_gsem_mergedPGS_AFR <- fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/gsem_fivefactor_ABCD_release_3.0_QCed.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_int_gsem = pgs_from_GSEM_INT_gwas_wsub.txt.gz,
         pgs_neuro_gsem = pgs_from_GSEM_NDD_gwas_wsub.txt.gz,
         pgs_sub_gsem = pgs_from_GSEM_SUB_gwas_wsub.txt.gz,
         pgs_comp_gsem = pgs_from_GSEM_COMP_gwas_wsub.txt.gz,
         pgs_psy_gsem = pgs_from_GSEM_PSY_gwas_wsub.txt.gz)

###Irwin GWAS 
abcd_IrwinPGS_EUR = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/irwin_gsem_ABCD_release_3.0_QCed.1kgCEU_maf0.01_rs_ref.pgs.merged") %>%
  rename(pgs_ext_irwin = pgs_from_ext_mvGWAS_finaledited.txt,
         pgs_int_irwin = pgs_from_int_mvGWAS_finalEdited.txt,
         pgs_ndd_irwin = pgs_from_ndd_mvGWAS_finalEdited.txt,
         pgs_tp_irwin = pgs_from_tp_mvGWAS_finalEdited.txt)
abcd_IrwinPGS_AFR = fread("/Users/quanfahe/Library/CloudStorage/OneDrive-UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/irwin_gsem_ABCD_release_3.0_QCed.1000G_AFR_QC_rsID.pgs.merged") %>%
  rename(pgs_ext_irwin = pgs_from_ext_mvGWAS_finaledited.txt,
         pgs_int_irwin = pgs_from_int_mvGWAS_finalEdited.txt,
         pgs_ndd_irwin = pgs_from_ndd_mvGWAS_finalEdited.txt,
         pgs_tp_irwin = pgs_from_tp_mvGWAS_finalEdited.txt)



###GSEM GWAS - Compulsive
#abcd_CompGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_CompulsiveGSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_comp_gsem = Pt_1)

###GSEM GWAS - Neurodevelopment
#abcd_NeuroGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_Neuro_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_neuro_gsem = Pt_1)

###GSEM GWAS - Thought
#abcd_THTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_THT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_tht_gsem = Pt_1)

###GSEM GWAS - Internalizing
#abcd_INTGSEMPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_INT_GSEM_12112021.all_score", sep = "", header = TRUE) %>%
#  rename(pgs_int_gsem = Pt_1)



###AUDIT
#abcd_AUDITPGS = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_AUDIT.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_audit = X1)

###ALCH Depend full discovery
#abcd_ALCHPGS_full = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ALCHdep_fulldisc.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchfull = X1)

###ALCH Depend eur discovery
#abcd_ALCHPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ALCHdep_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alcheur = X1)

###ALCH Depend afr discovery
#abcd_ALCHPGS_afr = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_ALCHdep_afr.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_alchafr = X1)

###CUD EUR GWAS
#abcd_CUDPGS_eur = read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Cross Race PGS Benchmark/ABCD PGS/ABCD_CUD_eur.all.score", sep = "", header = TRUE) %>%
#  rename(pgs_cudeur = X1)


##merge Race variable with PGS first; then rbind across race
abcd_race <- abcd_pdemo
abcd_race <- abcd_race %>%
  mutate(hispanic = ifelse(demo_ethn_v2==1 & demo_ethn_v2!=777 & demo_ethn_v2 !=999,1,0))
abcd_race <- abcd_race %>%
  mutate(Race = ifelse(demo_race_a_p___10==1 & hispanic == 0, 1, 
                       ifelse(demo_race_a_p___11==1 & hispanic == 0, 2, 
                              ifelse(hispanic==1, 3, 4))))

##white and hispanic and black subject ID
whiteHisp <- abcd_race$subjectkey[which(abcd_race$Race==1 | abcd_race$Race==3)]
black <- abcd_race$subjectkey[which(abcd_race$Race==2)]
##select white and hispanic from EUR reference PGSs 
abcd_indDisorder_whiteHispanic <- abcd_indDisorder_merged_EUR[abcd_indDisorder_merged_EUR$IID %in% whiteHisp,]
abcd_gsem_whiteHispanic <- abcd_gsem_mergedPGS_EUR[abcd_gsem_mergedPGS_EUR$IID %in% whiteHisp,]
abcd_extCons_whiteHispanic <- abcd_EXTconsPGS_EUR[abcd_EXTconsPGS_EUR$IID %in% whiteHisp,]
abcd_IrwinPGS_whiteHispanic <- abcd_IrwinPGS_EUR[abcd_IrwinPGS_EUR$IID %in% whiteHisp,]
##select black from AFR reference PGSs 
abcd_indDisorder_black <- abcd_indDisorder_merged_AFR[abcd_indDisorder_merged_AFR$IID %in% black,]
abcd_gsem_black <- abcd_gsem_mergedPGS_AFR[abcd_gsem_mergedPGS_AFR$IID %in% black,]
abcd_extCons_black <- abcd_EXTconsPGS_AFR[abcd_EXTconsPGS_AFR$IID %in% black,]
abcd_IrwinPGS_black <- abcd_IrwinPGS_AFR[abcd_IrwinPGS_EUR$IID %in% black,]

##merge using rbind 
abcd_indDisorder_threeGroup <- rbind(abcd_indDisorder_whiteHispanic, abcd_indDisorder_black)
abcd_gsem_threeGroup <- rbind(abcd_gsem_whiteHispanic, abcd_gsem_black)
abcd_extCons_threegroup <- rbind(abcd_extCons_whiteHispanic,abcd_extCons_black)
abcd_irwin_threegroup <- rbind(abcd_IrwinPGS_whiteHispanic,abcd_IrwinPGS_black)

###Merge all PGSs and change SUBJID
abcdPGSCombined <- Reduce(function(...) merge(..., all = TRUE), 
                      list(abcd_indDisorder_threeGroup,abcd_gsem_threeGroup,
                           abcd_extCons_threegroup,abcd_irwin_threegroup))

abcdPGSCombined <- abcdPGSCombined %>%
  rename(subjectkey = IID)
abcdPGSCombined$FID = NULL


correlationPGS<-corr.test(abcdPGSCombined[which(abcdPGSCombined$subjectkey%in%whiteHisp),-1])
corrplot::corrplot.mixed(correlationPGS$r)

correlationPGS_black<-corr.test(abcdPGSCombined[which(abcdPGSCombined$subjectkey%in%black),-1])
corrplot::corrplot.mixed(correlationPGS_black$r)

###CBCL
abcd_cbcls <- fread(paste0(abcdPath, "abcd_cbcls01.txt")) %>%
  slice(-1) %>%
  filter(eventname == "baseline_year_1_arm_1") %>%
  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))  ##---> event name is the same for all subject;

abcd_cbcl <- fread(paste0(abcdPath, "abcd_cbcl01.txt")) %>%
  slice(-1) %>%
  filter(eventname == "baseline_year_1_arm_1") %>%
  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))  ##---> event name is the same for all subject;

########################################dataset id has the same name but different number for each data frame 
########################################so it should not be used as a unique id. 
########################################I think it is the id for the measure as a whole
###KSADS 
abcd_pksads <- fread(paste0(abcdPath, "abcd_ksad01.txt")) %>%
  slice(-1) %>%
  filter(eventname == "baseline_year_1_arm_1") %>%
  select(-c("eventname","dataset_id","collection_title"))


####from KSADS, extract only the diagnoses
####then rename to short names for the disorders
######short name -> long name#####
#mdd - major depressive disorder; dysth - dysthymia (persistent depressive disorder); bip1, bip2 (bipolar I and bipolar II); 
#
#####ksads diagnosis extract####
####888 -> not administered due to branching logic from screening items
####555 -> not administered at this wave
abcd_pksadsDiag <- abcd_pksads %>%
  select(c("abcd_ksad01_id","subjectkey","src_subject_id","interview_date","interview_age","sex","collection_id",
           "ksads_1_840_p","ksads_1_841_p","ksads_1_842_p","ksads_1_843_p","ksads_1_844_p","ksads_1_845_p", 
           "ksads_2_830_p","ksads_2_831_p","ksads_2_832_p","ksads_2_833_p","ksads_2_834_p","ksads_2_835_p","ksads_2_836_p","ksads_2_837_p",
           "ksads_3_848_p","ksads_4_826_p","ksads_4_827_p","ksads_4_828_p","ksads_4_829_p","ksads_4_849_p","ksads_4_850_p",
           "ksads_5_857_p","ksads_5_858_p",	"ksads_6_859_p","ksads_6_860_p","ksads_7_861_p","ksads_7_862_p","ksads_8_863_p","ksads_8_864_p",
           "ksads_9_867_p","ksads_9_868_p","ksads_10_869_p","ksads_10_870_p","ksads_11_917_p","ksads_11_918_p",
           "ksads_12_925_p","ksads_12_926_p","ksads_12_927_p","ksads_12_928_p",
           "ksads_13_930_p","ksads_13_931_p","ksads_13_932_p","ksads_13_933_p","ksads_13_934_p","ksads_13_935_p",
           "ksads_13_936_p","ksads_13_937_p","ksads_13_938_p","ksads_13_939_p","ksads_13_940_p",
           "ksads_14_853_p","ksads_14_854_p","ksads_14_855_p","ksads_14_856_p",
           "ksads_15_901_p","ksads_15_902_p","ksads_16_897_p","ksads_16_898_p","ksads_16_899_p","ksads_16_900_p",
           "ksads_17_904_p","ksads_17_905_p", "ksads_18_903_p",
           "ksads_19_891_p","ksads_19_892_p","ksads_19_895_p","ksads_19_896_p","ksads_20_889_p","ksads_20_890_p",
           "ksads_21_921_p","ksads_21_922_p","ksads_22_970_p","ksads_22_969_p",
           "ksads_23_945_p","ksads_23_946_p","ksads_23_954_p","ksads_23_956_p","ksads_23_957_p","ksads_23_965_p",
           "ksads_24_967_p","ksads_24_968_p","ksads_25_865_p","ksads_25_866_p")) %>%
  rename(ksads_mdd_crt = ksads_1_840_p, ksads_mdd_rmi = ksads_1_841_p, ksads_mdd_pst = ksads_1_842_p,
         ksads_dysth_crt = ksads_1_843_p, ksads_dysth_rmi = ksads_1_844_p, ksads_dysth_pst = ksads_1_845_p,
         ksads_bip1man_crt = ksads_2_830_p, ksads_bip1dep_crt =  ksads_2_831_p, ksads_bip1hypoman_crt = ksads_2_832_p, 
         ksads_bip1man_pst = ksads_2_833_p, ksads_bip1dep_pst = ksads_2_834_p, ksads_bip2hypoman_crt = ksads_2_835_p, ksads_bip2dep_crt = ksads_2_836_p, ksads_bip2hypoman_pst = ksads_2_837_p,
         ksads_dmdd_crt = ksads_3_848_p,
         ksads_hallu_crt = ksads_4_826_p, ksads_hallu_pst = ksads_4_827_p,
         ksads_delu_crt = ksads_4_828_p, ksads_delu_pst = ksads_4_829_p,
         ksads_asspsy_crt = ksads_4_849_p, ksads_asspsy_pst = ksads_4_850_p, 
         ksads_pan_crt = ksads_5_857_p, ksads_pan_pst = ksads_5_858_p,
         ksads_agr_crt = ksads_6_859_p, ksads_agr_pst = ksads_6_860_p,
         ksads_sep_crt = ksads_7_861_p, ksads_sep_pst = ksads_7_862_p, 
         ksads_soc_crt = ksads_8_863_p, ksads_soc_pst = ksads_8_864_p, 
         ksads_pho_crt = ksads_9_867_p, ksads_pho_pst = ksads_9_868_p,
         ksads_gad_crt = ksads_10_869_p, ksads_gad_pst = ksads_10_870_p,
         ksads_ocd_crt = ksads_11_917_p, ksads_ocd_pst = ksads_11_918_p,
         ksads_enure_crt = ksads_12_925_p, ksads_enure_pst = ksads_12_926_p, 
         ksads_encop_crt = ksads_12_927_p, ksads_encop_pst = ksads_12_928_p, 
         ksads_an_binpur_rmi = ksads_13_930_p, ksads_an_binpur_pst = ksads_13_931_p, 
         ksads_an_rest_crt = ksads_13_932_p, ksads_an_rest_rmi = ksads_13_933_p, ksads_an_rest_pst = ksads_13_934_p, 
         ksads_bulim_crt = ksads_13_935_p, ksads_bulim_pst = ksads_13_936_p, ksads_bulim_rmi = ksads_13_937_p,
         ksads_binge_crt = ksads_13_938_p, ksads_binge_rmi = ksads_13_939_p, ksads_binge_pst = ksads_13_940_p, 
         ksads_adhd_crt = ksads_14_853_p, ksads_adhd_pst = ksads_14_854_p, ksads_adhd_rmi = ksads_14_855_p, ksads_adhd_unspe = ksads_14_856_p, 
         ksads_odd_crt = ksads_15_901_p, ksads_odd_pst = ksads_15_902_p, 
         ksads_cd_crt_child = ksads_16_897_p, ksads_cd_crt_adol = ksads_16_898_p, ksads_cd_pst_child = ksads_16_899_p, ksads_cd_pst_adol = ksads_16_900_p, 
         ksads_tic_unspe_crt = ksads_17_904_p, ksads_tic_unspe_pst = ksads_17_905_p, 
         ksads_asd_unspe = ksads_18_903_p,
         ksads_aud_crt = ksads_19_891_p, ksads_aud_pst = ksads_19_892_p, ksads_alc_unspe_crt = ksads_19_895_p, ksads_alc_unspe_pst = ksads_19_896_p,
         ksads_sud_crt = ksads_20_889_p, ksads_sud_pst = ksads_20_890_p,
         ksads_ptsd_crt = ksads_21_921_p, ksads_ptsd_pst = ksads_21_922_p,
         ksads_slp_crt = ksads_22_969_p, ksads_slp_pst = ksads_22_970_p, 
         ksads_sinjbeh_crt = ksads_23_945_p, ksads_sipass_crt = ksads_23_946_p, ksads_satt_crt = ksads_23_954_p,
         ksads_sinjbeh_pst = ksads_23_956_p, ksads_sipass_pst = ksads_23_957_p, ksads_satt_pst = ksads_23_965_p,
         ksads_hom_crt = ksads_24_967_p, ksads_hom_pst = ksads_24_968_p, 
         ksads_selmutism_crt = ksads_25_865_p, ksads_selmutism_pst = ksads_25_866_p) 

#replace 555 with NA
abcd_pksadsDiag <- abcd_pksadsDiag %>%
  na_if(555)
#originally was going to replace 888 with NA  -> but see Barch et al., pg. 10, 
##888 means the questions were not asked due to branching logic, based on screening items; so these 888 should be coded as "absent"
abcd_pksadsDiag <- abcd_pksadsDiag %>%
  mutate(across(starts_with("ksads"), ~recode(., '888' = '0')
                ))

###ABCLS -> only available for 2 year follow up
#abcd_abcls <- fread(paste0(abcdPath, "abcd_abcls01.txt")) %>%
#  slice(-1) %>%
#  filter(eventname == "2_year_follow_up_y_arm_1") %>%
#  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))

###parent demographic
abcd_pdemo <- fread(paste0(abcdPath, "pdem02.txt")) %>%
  slice(-1) %>%
  filter(eventname == "baseline_year_1_arm_1") %>%
  select(-c("eventname","dataset_id","study_cohort_name","collection_title","interview_date","collection_id","src_subject_id","sex")) %>%
  select("subjectkey","demo_prim","demo_brthdat_v2","demo_ed_v2","demo_sex_v2","interview_age",
         starts_with("demo_race_a_p"),"demo_ethn_v2","demo_ethn2_v2",
         "demo_prnt_age_v2","demo_prnt_gender_id_v2","demo_prnt_marital_v2","demo_prnt_ed_v2",
         "demo_prnt_empl_v2","demo_prnt_empl_time","demo_prnt_income_v2","demo_prnt_prtnr_v2",
         "demo_prnt_prtnr_bio","demo_prtnr_ed_v2","demo_prtnr_empl_v2","demo_prtnr_empl_time",
         "demo_prtnr_income_v2","demo_comb_income_v2",starts_with("demo_fam_exp"))
         
###medical history and service utilization --> only at 1 year follow up and 2 year follow up
#abcd_medServ <- fread(paste0(abcdPath, "abcd_lpmh01.txt")) %>%
#  slice(-1) %>%
#  filter(eventname == "2_year_follow_up_y_arm_1") %>%
#  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))

##anthropometric measure
#abcd_anthro <- fread(paste0(abcdPath, "abcd_ant01.txt")) %>%
#  slice(-1) %>%
#  filter(eventname == "baseline_year_1_arm_1") %>%
#  select(-c("eventname","dataset_id","study_cohort_name","collection_title"))

rm(list = c("abcd_pksads"))

##merge files into one dataframe
#abcd_merged <- abcd_cbcl %>%
#  full_join(abcd_pksadsDiag, by = mergeCol) %>%
#  full_join(abcd_anthro,by = mergeCol) %>%
#  full_join(abcd_pdemo,by = mergeCol)
#abcd_merged %>%
#  glimpse()

##make numeric data into numeric
#abcd_merged[,c(6, 8:271)] <- abcd_merged[,c(6, 8:271)] %>%
#  lapply(as.numeric) %>%
#  as.data.frame()

##recode sex into 0 (M), and 1 (F) 
#abcd_merged$sex <- recode(abcd_merged$sex, 'M' = 0, 'F' = 1)

####------------------Add Health----------------####
AddHealthAll = fread("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/FYP/Analysis/Datafile/w1-w4Raw.csv")
AddHealthFull <- read.spss("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Add Health related/reorganize.july2021/PrimaryData.processed.7.26.2021.sav",
                           to.data.frame = TRUE)
AddHealthselected <- c("asbsub1",PC, names(AddHealthFull[1:8]), "CALCAGE1", "ADHD_totalscore","inattention","hyperactivity","COVSUM1","OVTSUM1","PSUSCID","REGION")
AddHealthFull <- AddHealthFull[,AddHealthselected]
##add Depression score at Wave 1 
MDD1 = paste("H1FS",c(6,1,15,16,17,3,7,4,5,11), sep ="")
AddHealthFull <- AddHealthAll %>%
  select(all_of(MDD1),"AID") %>%
  merge(AddHealthFull, by = "AID")
##Add PGSs computed by Q
AddHealthPGS <- read.delim("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/Genetic processing/Q lab/AddHealth_set1_traits_other.txt")

selectPGSAddHealth <- data.frame(AddHealthPGS[,c("FID","IID","D0004","D0008","D0009","D0011","D0013","D0020","D0021","D0024","D0025","D0028",
                                                 "D0029","D0030","D0036","D0039","D0040","D0041","D0044","D0046","D0048","D0051","D0054","D0077",
                                                 "D0094","D0096","D0105")])

##Link data file
Link = read.table("/Users/quanfahe/OneDrive - UW-Madison/GradSchool/Research/FYP/Analysis/GSEM/My replication/AddHealth dbGap ID Link File.dat",
                  header = TRUE)
tempAddHealthPGSLinked <- merge(Link[,c("AID", "GID")], selectPGSAddHealth, by.x = "GID", by.y = "IID")
###merge PGS with Phenotypes
AddHealthFull <- merge(AddHealthFull,tempAddHealthPGSLinked)

rm(list = c("AddHealthAll","AddHealthPGS","AddHealthASB"))

##rename the PGSs
AddHealthFull <- AddHealthFull %>%
  rename(SCZ_PGS2014 = "D0004",
         PTSD_DuncanEA = "D0008",
         OCD_PGC2017 = "D0009",
         MDD_PGC2018 = "D0011",
         Insom_Hammerschlag_all = "D0013",
         BIP_PGC = "D0020",
         ASD_PGC17EA = "D0021",
         ANX_ANGST_FS = "D0024",
         AN_PGC2017 = "D0025",
         ASD_PGC2019 = "D0028",
         BIP_PGC2018 = "D0029",
         TS_Yu2018 = "D0030",
         SCZ_Pardinas = "D0036",
         ANX_iPSYCH1 = "D0039",
         ANX_iPSYCH2 = "D0040",
         ANX_iPSYCH3 = "D0041",
         CROSS_iPSYCH2019 = "D0044",
         AN_PGC2019 = "D0046",
         PTSD_PGC2019_all = "D0048",
         PTSD_PGC2019_EUR = "D0051",
         PTSD_PGC2019_AA = "D0054",
         MDD_Howard = "D0077",
         Insom_Jansen = "D0094",
         ASD_Matoba2020EUR = "D0096",
         COVID = "D0105")
names(AddHealthFull)
